chore: move list_active_job_ids to client api
