+++ test/unit/deadline_job_attachments/bucket_sweeper/test_job_attachments_sweeper.py
@@ -1,14 +1,11 @@
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 
 from botocore.exceptions import BotoCoreError
-from deadline.client.api._list_jobs_by_filter_expression import JobFetchFailure
-from deadline.client.exceptions import DeadlineOperationError
 import pytest
 import os
 import csv
 from typing import Dict, List, Any
 from pathlib import Path
-from datetime import datetime, timezone
 from unittest.mock import Mock
 from deadline.job_attachments.bucket_sweeper.job_attachments_sweeper import JobAttachmentsSweeper
 from deadline.job_attachments.exceptions import (
@@ -16,10 +13,6 @@ from deadline.job_attachments.exceptions import (
     JobAttachmentS3BotoCoreError,
 )
 
-@pytest.fixture
-def mock_boto3_session():
-    """Fixture to create mock AWS boto3 session"""
-    return Mock()
 
 @pytest.fixture
 def mock_s3() -> Mock:
@@ -37,10 +30,9 @@ def mock_deadline() -> Mock:
     return Mock()
 
 @pytest.fixture
-def processor(mock_boto3_session, mock_s3, mock_s3_control, mock_deadline) -> JobAttachmentsSweeper:
+def processor(mock_s3, mock_s3_control, mock_deadline) -> JobAttachmentsSweeper:
     """Fixture to create JobAttachmentsSweeper instance with mock clients"""
     return JobAttachmentsSweeper(
-        boto3_session=mock_boto3_session,
         s3_client=mock_s3,
         s3_control_client=mock_s3_control,
         deadline_client=mock_deadline,
@@ -61,94 +53,6 @@ def test_dir(tmp_path: Path) -> Path:
 
 
 class TestJobAttachmentsSweeper:
-
-    def test_get_active_job_ids_empty_queue_list(self, processor):
-        """Test behavior with empty queue list."""
-        result: Dict[str, List[str]] = processor._get_active_job_ids([], datetime.now(timezone.utc))
-        assert result == {}
-
-    def test_get_active_job_ids_empty_jobs_response(
-        self, processor: JobAttachmentsSweeper, monkeypatch: pytest.MonkeyPatch
-    ):
-        """Test behavior when no jobs are returned."""
-
-        def mock_list_jobs_by_filter_expression(*args, **kwargs):
-            return []
-
-        monkeypatch.setattr(
-            f"{processor.__module__}._list_jobs_by_filter_expression",
-            mock_list_jobs_by_filter_expression,
-        )
-
-        result: Dict[str, List[str]] = processor._get_active_job_ids(
-            ["queue-1"], datetime.now(timezone.utc)
-        )
-        assert result == {"queue-1": []}
-
-    def test_get_active_job_ids_deadline_error(
-        self, processor: JobAttachmentsSweeper, monkeypatch: pytest.MonkeyPatch
-    ):
-        """Test get active job ids when listing function fails."""
-
-        # Setup mock function response
-        def mock_list_jobs_by_filter_expression(*args, **kwargs):
-            raise DeadlineOperationError()
-
-        monkeypatch.setattr(
-            f"{processor.__module__}._list_jobs_by_filter_expression",
-            mock_list_jobs_by_filter_expression,
-        )
-
-        with pytest.raises(JobAttachmentsSweeperError):
-            processor._get_active_job_ids(
-                ["queue-1"], retention_datetime=datetime.now(timezone.utc)
-            )
-
-    def test_get_active_job_ids_job_fetch_failure_error(
-        self, processor: JobAttachmentsSweeper, monkeypatch: pytest.MonkeyPatch
-    ):
-        """Test get active job ids when listing function fails."""
-
-        # Setup mock function response
-        def mock_list_jobs_by_filter_expression(*args, **kwargs):
-            raise JobFetchFailure()
-
-        monkeypatch.setattr(
-            f"{processor.__module__}._list_jobs_by_filter_expression",
-            mock_list_jobs_by_filter_expression,
-        )
-
-        with pytest.raises(JobAttachmentsSweeperError):
-            processor._get_active_job_ids(
-                ["queue-1"], retention_datetime=datetime.now(timezone.utc)
-            )
-
-    def test_get_active_job_ids(self, processor: JobAttachmentsSweeper, monkeypatch: pytest.MonkeyPatch):
-        """Test getting job ids for multiple queues."""
-
-        # Setup mock function response
-        def mock_list_jobs_by_filter_expression(*args, **kwargs):
-            return [{"jobId": "job-1"}, {"jobId": "job-2"}]
-
-        monkeypatch.setattr(
-            f"{processor.__module__}._list_jobs_by_filter_expression",
-            mock_list_jobs_by_filter_expression,
-        )
-
-        # Call method
-        queue_ids: List[str] = ["queue-1", "queue-2", "queue-3"]
-        retention_datetime: datetime = datetime.now(timezone.utc)
-        result: Dict[str, List[str]] = processor._get_active_job_ids(
-            queue_ids, retention_datetime=retention_datetime
-        )
-
-        expected_result: Dict[str, List[str]] = {
-            "queue-1": ["job-1", "job-2"],
-            "queue-2": ["job-1", "job-2"],
-            "queue-3": ["job-1", "job-2"],
-        }
-        assert list(result.keys()) == list(expected_result.keys())
-        assert list(result.values()) == list(expected_result.values())
 
     def test_create_tag_manifest_empty_list(self, processor: JobAttachmentsSweeper, test_dir: Path):
         """Test creating a tag manifest with an empty delete list."""
